pyproject.toml
[project]
name = "chromag"
version = "0.1.0"
authors = [
  { name="Michael Galloy", email="mgalloy@ucar.edu" },
]
description = "Calibrate, analyze, and browse ChroMag data"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy >= 1.20.2",
  "astropy >= 3.2.2",
  "epochs >= 0.2.0",
  "python-dateutil >= 2.8.1",
]

[project.optional-dependencies]
dev = [
    "pylint",
    "black",
    "pre-commit",
    "pytest-cov",
    "Sphinx",
    "sphinx_rtd_theme",
]

[project.urls]
"Homepage" = "https://github.com/NCAR/chromag-pipeline"
"Bug Tracker" = "https://github.com/NCAR/chromag-pipeline/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
chromag = "chromag.cli:main"

[tool.pytest.ini_options]
# increment the --cov-fail-under as we increase test coverage
addopts = "--cov --cov-report html:coverage_html --cov-report term-missing --cov-fail-under 3"

[tool.coverage.run]
source = ["chromag"]

[tool.pylint]
max-line-length = 90
# disable = [
#     "C0103", # (invalid-name)
#     "C0114", # (missing-module-docstring)
#     "C0115", # (missing-class-docstring)
#     "C0116", # (missing-function-docstring)
#     "R0903", # (too-few-public-methods)
#     "R0913", # (too-many-arguments)
#     "W0105", # (pointless-string-statement)
# ]